Moved 3 chords: Escape, " and CpsLock
Moved "Escape" to LLLL, because “Escape" is a dangerous key, as it can
get out of composition of a text and result in the loss of a lot of
text.
Moved chord " into old "Escape" (LOOL) as it is very often used chord.
Moved "CpsLock" to LOLL as this key is often used and the only "Lock"
key present on a Mac keyboard.