Added remaining NUM versions of all chords
This allows NUM key to be used as a full Ctrl key on Mac without having
to discover which key will I need and having to add it.
This will also allow combinations like Ctrl+Shift+4 (pressed as NUM+$)
to work using only one hand, even though NUM+SHIFT is not possible to
press together using only one hand.
The only downside of this is you need to remember that Shift+4 = "$",
but the "Cheat sheet" for this info is any standard keyboard.

Also fixed “`” and added “Es”