Experimental: Moved <Space> from MOOO to LOOO.
I noticed that when I press the <Space> key after "ph", "ing" and other
MCCs, I often do not see the space,
so I have to artificially wait till I press <Space>, which slows me
down significantly.
I decided to move <Space> from MOOO to LOOO.
I had to move <Tab> to SHIFT+MOOO. I have not found Tab useful enough
to require a dedicated button.
I added CTRL+SHIFT+MOOO as Shift+<Tab>.
I moved <Delete> to SHIFT+ROOO.
It might be a good adea to not use MOOO chord at all as it might reduce
error rate,
because autocorrection will be easier to do.
E.g., when a non-space button in "c" chord is erroneously not pressed,
a latter will be missing instead of a Space inserted in the middle of
the word.