Added "I" (capital i) as MOOO.
MOOO key should be used by itself as a letter with a high frequency,
but the one that is not encountered in the middle of the word.  "I"
(capital i) is probably the best choice.
This will allow seeing mistyped chords involving MOOO key  and help
with correcting typos.
Plus typing will be a little faster as I will have one more
easy-to-type letter.