More simplifications of chords
Swapped "ion" and "b" as "b" is used 4 times more than "ion" and OLOM
only moves middle finger away from the home row.
Moved "-" to MOMM, "0"(zero) to MMRO, "it" to MRMO, "st" to OLOR as I
hear that I'm not able to press buttons "s" and "t" at the same time.
Added <Del> as RRRO.