Introducing Backspice2 layout
I realized that only 7 MCCs don't match corresponding letters:
"and", "ed", "et", "ing", "ion", "ou", "st"
I have to let go of "ion", as it's not possible to "fix" this MCC.
"et" has to go too as both letters are on the middle row and "et" has
lower frequency.

Move "?" to ORMM, ":" to MLMO, <"> to RMOM
Move "d" to MOOR, "and" to MORR(and), "ed" to MMOR(ed)
Move "g" on OROR, "ing" stays on ORRR(ing)
Move "u" to MOLO, "ou" to MLLO(ou),
Move "q" to MMLO, keeping the relationship to "u".
MMOO/MROO cold be used for Left/Right, OMMO as Up and OMOM as Down
arrows.
Moving ARROWS allows fixing "st".
Move "p" to OROM, "-" to MMOM, "w" to LOOM, "y" to LMMO, "k" to LMOM
Move "<" to LRRO, ">" to LMMO
Move "{" to LORO and "}" LOLO
Move all numbers to be more frequency-based-friendly